chore(growth): move product push ad logic into a kea logic
Pulls the derived state (product info, display, flag gating, visibility) and
the impression/click/dismiss analytics out of NavPanelProductPushAd and into
navPanelProductPushAdLogic, so the component no longer owns business logic or a
side-effecting useEffect. The display map moves to its own module to keep the
component/logic import graph acyclic.
